Moved: reddit_data -> reddit_data_interface
The .gitignore was aimed at a directory full of code.

Move any data files that need to be loaded to a dir called 'data'
