+++ process_trees.py
@@ -8,7 +8,7 @@ from torchtext.data.utils import get_tokenizer
 from sklearn.feature_extraction.text import CountVectorizer
 from nltk.corpus import stopwords
 #
-from reddit_data import RedditTreeBuilder, JsonDataSource
+from reddit_data_interface import RedditTreeBuilder, JsonDataSource
 from machine_learning.glove_embedding import UNKNOWN_WORD, glove, word2idx
 from text_preprocessing import stemming_tokenizer, get_parent_indices
 
@@ -162,7 +162,7 @@ def print_hate_speech(trees):
 
 if __name__ == "__main__":
     # Load reddit data
-    path_to_json_data = "reddit_data/RC_2006-12" # for example, use comments from 2006
+    path_to_json_data = "data/RC_2006-12"  # for example, use comments from 2006
     jds = JsonDataSource(path_to_json_data)
     rt = RedditTreeBuilder(jds)
