--- reddit_data/data_sources/__init__.py
@@ -1,1 +1,0 @@
-from reddit_data.data_sources.json_data_source import JsonDataSource
