+++ reddit_data_interface/data_sources/__init__.py
@@ -1,0 +1,1 @@
+from reddit_data_interface.data_sources.json_data_source import JsonDataSource
