+++ reddit_data_interface/data_sources/data_source.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Collection, Iterable
 
-from reddit_data.data_models import RedditComment
+from reddit_data_interface.data_models import RedditComment
 
 
 class DataSource(ABC):
